requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
cohere==5.2.4
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
pydantic==2.12.5
pydantic-settings==2.12.0
